STAR-121 Port SAI from DSE to OS
STAR-348 moved postings decoded to mark on postings reader close (#87)

fixed up PostingsTest and others

removed new line

(cherry picked from commit fbdbec306ebd383aa3ac05236e93c21ce869cbce)
(cherry picked from commit fabfca15c68c05ae9a5903767ca150db03b1a2d5)
(cherry picked from commit cf7714e5be99b2ec77d34b44b4aff551516d7d09)
(cherry picked from commit aea6839269dde72b2ee586e4ee4728e271773d4d)
(cherry picked from commit 520d1fa339af36d423aeaa0a6e2eb42b014f3fd1)
(cherry picked from commit e5a1da626247621412b33250ef6251992f38dfce)
(cherry picked from commit 3056f403047412235e13f16aced385dcaef4ab89) (+1 squashed commit)
Squashed commits:
[7db49104811] STAR-360: Change compaction pause to be more specific to rewrite

(cherry picked from commit 16b3e35db2a8d6c66bd5fead3c1240116a3ae8c4)
(cherry picked from commit b913058c6cfe47624ced944255877630eb8e3c3f)
(cherry picked from commit 641a5cfaa403fc60ec45110518ebc30810b5eec6)
(cherry picked from commit 43bbcfecb5b7fdb83d6f0f0ce8986e2ea64d3250)
(cherry picked from commit 6f80df980ff5e45b9ceaf2d0fee3edb25732ef2c)
(cherry picked from commit 745de8275e7f40068a0331fdeebabea1a5498d19)
(cherry picked from commit 96cd4544c84bb161e26b18c4f362e49ca28ab7a1) (+10 squashed commits)
Squashed commits:
[1ea07eff83f] STAR-361: Query all ranges at once for SAI distributed queries (#85)

(cherry picked from commit 5ba3c6bd57ea437e670c9f3734d36cd5f75822db)
(cherry picked from commit c7e186dd5648ac96472c1bf519d35082c37d6d33)
(cherry picked from commit efc66d1c1f7a729ad14a359963d968290286b88c)
(cherry picked from commit 7d5630f14aa4f915845e4f5738fab899477fa158)
(cherry picked from commit 9e684651ddb72012692210fd04b03bb7b3f9161d)
(cherry picked from commit 57103ba49a9f6d55d7b8df4e9102f837454f9da4)
(cherry picked from commit 1ae42ea6e31b15bc0b7bf15c7aeefdd91768497c)
[e0804e871a1] STAR-347: Fix FailureTest failure

(cherry picked from commit f090eb6dce1e0e425a3d0ba209f8cb42e0ac540a)
(cherry picked from commit ea105ce23fb105b0383c485617b5ddcec81a5a9d)
(cherry picked from commit 8d187208d907bed215fc5baf5814f2125b540fc0)
(cherry picked from commit 3c31759f010106c01413ab99152cd244f331b46d)
(cherry picked from commit 88ee842b6508882870ad7ec45d429154041ec18f)
(cherry picked from commit db0ab05488bd57f1bff0e3c0fe2c9499e58c6b96)
(cherry picked from commit 5f2c269ce033757fdae3cd628324fb17e5feaa3b)
[1229d4c1640] STAR-121: Storage Attached Index implementation and testing

(cherry picked from commit 430c24026fb2a4afea735475408f0ac34944478f)
(cherry picked from commit acbfeb80cf865ebe231146ad8ed54e03fef06cd3)
(cherry picked from commit 54af59d4f31bff6ee7cbf8f670e7d82e94cc00d2)
(cherry picked from commit 30425c49353bb88e756f1ea33cdb50d8932332b4)
(cherry picked from commit 52d50e8ebfea9c6ee3cec6f170a1af26dd9a4310)
(cherry picked from commit 94773ba88c6c7cf13e9f859fd75aef2d77f44cab)
(cherry picked from commit ac2f66c401a45d0437198b6139efab74260c4ba5)
[affe267a4ac] STAR-121: Test support for SAI

(cherry picked from commit 6501503cf75a9a1d2f0ace5eb5d56bb9f28bb361)
(cherry picked from commit f4d12eae47d3585897798df1b437bb4e4177391b)
(cherry picked from commit c1cad34f418e4dec97bc02feb373980a1d6cd485)
(cherry picked from commit 27ac10a37e9a7879baeb355b9a4cf5660f61a802)
(cherry picked from commit 47becde9af455009334515e3e6032d3683e30838)
(cherry picked from commit dc715880c196af11e2c4ab9206989aa75ff6e167)
(cherry picked from commit d35ed8291f0a74fc2f21e04fb0ce11f9c7c3d263)
[ad7cf475338] STAR-121: Key iteration and utility support for SAI

(cherry picked from commit 36c018bca0a27c53c9c91a40bf0621e56fc16152)
(cherry picked from commit daf5f712e38ca71357112d496ecebb370b47d90e)
(cherry picked from commit 868c0a730f93a1b1f59c7c0b08345183fc4add3d)
(cherry picked from commit 35f3d50cb868829c4edbac9553a355ff5e3561d3)
(cherry picked from commit 97e0bf6e93c55b800803833a317d1583c65bfe9b)
(cherry picked from commit 3dc326a2dc37906594892f9895fcdc1872c68d1c)
(cherry picked from commit 2e7f9ed23ea03e5e2d068998262538933279b4e9)
[c9d2d226334] STAR-121: On-disk component support for SAI

(cherry picked from commit 89668cf917f0ea6a663f88e3081865644daae150)
(cherry picked from commit 906845012771c0f10d6e051138883915f592ab8f)
(cherry picked from commit c80316cada90998a0f558b582425c2ac40899d40)
(cherry picked from commit 39f03c68628b9a5b7e584fbfea34bac67677f03b)
(cherry picked from commit 3ebeb25918325894a05032b2b96bcde9482982be)
(cherry picked from commit 7fd43c04c63caf932bddbdf42d001b1852027f18)
(cherry picked from commit b3481de70372eb4e53fe051d7961311942f480d5)
[6a3d288056a] STAR-121: Config support for SAI

(cherry picked from commit 5db0b08fc6eee05fe914261600de51c1524685d0)
(cherry picked from commit 24192212f5b456ee16aed5a4c065da43b393c46c)
(cherry picked from commit 36d022e3d553f26769a7f7a7aa6fc067026e8799)
(cherry picked from commit 37dd663047408b2a4d957d2f82ee6b03ae47212d)
(cherry picked from commit 8af888ff6f402c8962c10bbe888878a99f32e178)
(cherry picked from commit 7721e9b3aff762da6ebdd818420480e158ff7b13)
(cherry picked from commit 7df7779fdf5b1af99cab15e7e2ecc8cf08b5d118)
[bdaa00f65d4] STAR-121: Build dependencies for SAI

(cherry picked from commit 32580250852ce46f2f71e0631241eb64e51753a8)
(cherry picked from commit 77850e3a2d37db82a6f0b52aa420bf3970266dbb)
(cherry picked from commit ee4575d13405ea40443b3ce3fa68ed76d7b00024)
(cherry picked from commit 65e02b3a08844fff0aabfabe5103fc8b91d039bc)
(cherry picked from commit dfd29dc4021deb64546caf4049b7ab7d401f5470)
(cherry picked from commit c74939d2cd129041cf58c61e6af73af424a1038a)
(cherry picked from commit 13fde650cbe142ea35ff769b0b1145d9f37802b9)
[3e97eb5a4b7] STAR-121: Add index support to select statement restrictions

(cherry picked from commit abe04b3f031d01f543eb8f4a1f0c3665e710e96f)
(cherry picked from commit 0249d0559e3015cdbc232d05e4c88e22ee349c61)
(cherry picked from commit bbc9ca4222bf7a2e57c7ab4ec777ce6ad0e49542)
(cherry picked from commit 925a312b4627f7e41b1d8a2c7e9073be578fe1e6)
(cherry picked from commit fa021cdca7ed7ce62df2dd2e9cfba9ebde5d6b52)
(cherry picked from commit 3540945e1b6e91abc4867bdda56593c09c863548)
(cherry picked from commit d39b35be49eb68467ce1b2de528bb90732caac34)
[51c216c82fe] STAR-121: Index group API changes

(cherry picked from commit eec063120d7c617a5bc7215234cbaab60d8b78bc)
(cherry picked from commit 8363f8e4696741f0a31e072921e20344e7e5686b)
(cherry picked from commit 20b4628d0de65195f00d6752fb3a1244cb2411a8)
(cherry picked from commit f84f69453936b6047c6125c5f46cfe8c7a15241b)
(cherry picked from commit a5df8862facfc3f650a02c37560992b9b516335b)
(cherry picked from commit 1041d46eb22ba936440b818aeebb61a8e68a734c)
(cherry picked from commit defc28e1be12a9cb430d23103f6c86fb31135687)
(cherry picked from commit c6ed9b144713af5fe82dfb850c16f6af97719604)
(cherry picked from commit aaa1ef37fdcf14ea317c612f5e29f2e6fb9d234d)

+++ src/java/org/apache/cassandra/index/IndexRegistry.java
@@ -33,6 +33,7 @@ import javax.annotation.Nullable;
 import org.apache.cassandra.config.DatabaseDescriptor;
 import org.apache.cassandra.cql3.Operator;
 import org.apache.cassandra.db.*;
+import org.apache.cassandra.db.compaction.OperationType;
 import org.apache.cassandra.db.filter.RowFilter;
 import org.apache.cassandra.db.lifecycle.LifecycleNewTracker;
 import org.apache.cassandra.db.marshal.AbstractType;
@@ -43,7 +44,7 @@ import org.apache.cassandra.exceptions.InvalidRequestException;
 import org.apache.cassandra.index.transactions.IndexTransaction;
 import org.apache.cassandra.io.sstable.Component;
 import org.apache.cassandra.io.sstable.Descriptor;
-import org.apache.cassandra.io.sstable.SSTableFlushObserver;
+import org.apache.cassandra.io.sstable.format.SSTableFlushObserver;
 import org.apache.cassandra.schema.ColumnMetadata;
 import org.apache.cassandra.schema.IndexMetadata;
 import org.apache.cassandra.schema.TableMetadata;
@@ -228,7 +229,7 @@ public interface IndexRegistry
 
             @Nullable
             @Override
-            public Index.Indexer indexerFor(Predicate<Index> indexSelector, DecoratedKey key, RegularAndStaticColumns columns, long nowInSec, WriteContext ctx, IndexTransaction.Type transactionType, Memtable memtable)
+            public Index.Indexer indexerFor(Predicate<Index> indexSelector, DecoratedKey key, RegularAndStaticColumns columns, int nowInSec, WriteContext ctx, IndexTransaction.Type transactionType, Memtable memtable)
             {
                 return null;
             }
